requirements.txt
streamlit==1.24.0
langchain
pypdf
chromadb